renamed dirs and files
