+++ .gitignore
@@ -144,7 +144,7 @@ dmypy.json
 # Pyre type checker
 .pyre/
 
-# pytype static type analyzer
+# pytype static type analyzers
 .pytype/
 
 # Cython debug symbols
