+++ hypex/analyzers/analyzer.py
@@ -2,7 +2,7 @@ import warnings
 from abc import ABC
 
 from hypex.dataset.dataset import Dataset
-from hypex.experiment.experiment import ComplexExecutor
+from hypex.experiments.base import ComplexExecutor
 
 
 class Analyzer(ComplexExecutor, ABC):
