+++ hypex/transformers/transformers.py
@@ -2,7 +2,7 @@ from typing import Any, Union
 
 from hypex.dataset.dataset import Dataset
 from hypex.dataset.dataset import ExperimentData
-from hypex.experiment.experiment import Executor
+from hypex.experiments.base import Executor
 
 
 class Shuffle(Executor):
